% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_location.R
\docType{data}
\name{admin_location}
\alias{admin_location}
\title{Admin1 location look_up table}
\format{
A data frame with 3320 rows and 5 variables:
\describe{
\item{ISO}{ISO 3166-1 alpha-3 country code}
\item{Country}{country name}
\item{admin_name}{name of administrative division}
\item{long}{longitude of admin1}
\item{lat}{latitude of admin1}
}
}
\usage{
admin_location
}
\description{
This is a dataset with longitude and latitude of admin1
}
\references{
\url{https://gadm.org/download_world.html}
}
\keyword{datasets}
