% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness_report.R
\name{check_completeness}
\alias{check_completeness}
\title{Prints a completeness report for a country or admin1 unit}
\usage{
check_completeness(hit_data, country, admin1)
}
\arguments{
\item{hit_data}{the full HIT-COVID database pulled from GitHub, pulled using \link{hit_pull}}

\item{country}{country ISO 3166-1 alpha-3 code}

\item{admin1}{vector of the first administrative unit GID codes to filter the data to
(see \link{geo_lookup} for concordance of admin 1 codes to names).}
}
\description{
The admin1 units that are included in the HIT-COVID database can be found by using \link{get_admin1}
}
