% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookup.R
\docType{data}
\name{geo_lookup}
\alias{geo_lookup}
\title{Country and Admin1 Code Look-up Table}
\format{
A data frame with 3611 rows and 5 variables:
\describe{
\item{country}{ISO 3166-1 alpha-3 country code}
\item{admin1}{first administrative unit code (following GADM5 unless otherwise noted)}
\item{country_name}{country name}
\item{admin1_name}{level 1 administrative unit name}
\item{continent}{continent name}
\item{alpha_2}{ISO 3166-1 alpha-2 country code}
\item{admin1_ISO}{ISO 3166-2 admin1 code for selected countries}
\item{ons_region_code_uk}{Region code unique to the UK}
}
}
\usage{
geo_lookup
}
\description{
A dataset with the country codes and admin1 codes with their corresponding names.
It can be used to identify codes to use when filtering the HIT-COVID database.
}
\references{
\url{https://gadm.org/}
}
\keyword{datasets}
