% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_filter.R
\name{get_admin1}
\alias{get_admin1}
\title{Lists the admin1 codes for a country}
\usage{
get_admin1(hit_data, country)
}
\arguments{
\item{hit_data}{the full HIT-COVID database pulled from GitHub, pulled using \link{hit_pull}}

\item{country}{country ISO 3166-1 alpha-3 code}
}
\description{
Prints the admin1 levels for a given country that have data in the HIT-COVID database. The database
should first be pulled using \link{hit_pull} and fed into this function.
}
