% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_case.R
\name{get_first_case}
\alias{get_first_case}
\title{Get date of first case and first death}
\usage{
get_first_case(source = c("WHO", "ECDC"))
}
\arguments{
\item{source}{the source of the case data that is used to determine the date of first case and first
death: one of "ECDC" or "WHO" (default is "WHO")}
}
\value{
A dataframe with three columns:
\enumerate{
\item \code{country} the country code to link with the HIT-COVID database
\item \code{first_case} the date of the first case of COVID-19 in that country according to
the source specified.
\item \code{first_death} the date of the first death in that country from COVID-19 according
to the source specified.
}
}
\description{
Uses the \code{covidregionaldata} package to find the date of the first case of COVID-19 and the
date of the first death from COVID-19. The source is either the European Centre for Disease
Control (ECDC) or the World Health Organization (WHO).
}
\details{
This function can be run on its own to get the first case and death information if that is of
interest. However, it is primarily meant to be run inside of \link{hit_pull} to add this
information to the HIT-COVID database.
}
\examples{

\donttest{
#Pulling from ECDC
firsts <- get_first_case(source = "ECDC")

#Pulling from WHO
firsts <- get_first_case(source = "WHO")
}

}
\references{
Sam Abbott, Katharine Sherratt, Jonnie Bevan, Hamish Gibbs, Joel Hellewell, James Munday,
Paul Campbell and Sebastian Funk (2020). covidregionaldata: Subnational Data for the
Covid-19 Outbreak. R package version 0.6.0.
https://CRAN.R-project.org/package=covidregionaldata
}
\seealso{
\link{hit_filter}, \link{get_national_data}
}
