% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_filter.R
\name{get_interventions}
\alias{get_interventions}
\title{Lists the intervention_group codes}
\usage{
get_interventions()
}
\description{
Prints the valid values of the \code{intervention_group} argument used in \link{hit_filter}
and therefore also in \link{intervention_timeline} and \link{intervention_map}.
}
\seealso{
\link{hit_filter}, \link{intervention_timeline}, \link{intervention_map}
}
