% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_filter.R
\name{hit_pull}
\alias{hit_pull}
\title{Pulls HIT-COVID database from GitHub}
\usage{
hit_pull(add_first_case = TRUE)
}
\arguments{
\item{add_first_case}{a logical indicating if information about the first case and the first
death for each country should be added to the dataset}
}
\value{
A dataframe with columns as described in the GitHub README
(\url{https://github.com/HopkinsIDD/hit-covid})
}
\description{
This function pulls the HIT-COVID database from the GitHub repo:
\url{https://github.com/HopkinsIDD/hit-covid}.
}
\details{
ADD DETAILS OF FIRST CASE/FIRST DEATH INFO
}
\examples{

}
\seealso{
\link{hit_filter}
}
