% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_epi.R
\name{intervention_epi}
\alias{intervention_epi}
\title{Plots case counts with intervention status timelines}
\usage{
intervention_epi(
  hit_data,
  country,
  source = c("WHO", "ECDC"),
  case_threshold = 0,
  first_date = NULL,
  last_date = NULL,
  date_format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{hit_data}{the full HIT-COVID database pulled from GitHub, pulled using \link{hit_pull}}

\item{country}{vector of ISO 3166-1 alpha-3 country codes to filter the data to
(see \link{geo_lookup} for concordance of country codes to names)}

\item{source}{the source of the daily case count data, one of "ECDC" or "WHO" (default is "WHO").}

\item{case_threshold}{threshold the total number of cases needs to exceed to start the plot
(default is 0).}

\item{first_date}{character string indicating the earliest date to plot (default is first day where
the total number of cases is greater than the \code{case_threshold}).}

\item{last_date}{character string indicating the latest date to plot (default is today)}

\item{date_format}{character string indicating the format of the date inputs (default is "\%m/\%d/\%Y").}
}
\description{
This function plots the daily case counts for a given country from either the WHO or ECDC with
as seven day rolling average. The status of select intervention information (border closures,
household confinement, universal mask mandates, restaurant closures, primary school closures,
and retail store closures) are also plotted.
}
\details{
The HIT-COVID database must first be loaded with \link{hit_pull} and fed into this function.
Only one country can be plotted at a time and the case counts for this country will be plotted
with select intervention data. The user can specify the date range for the plot using
\code{first_date} and \code{last_date}. Another way to restrict the plot is to change
\code{case_threshold} then the earliest date of the plot will be when the total number of cases
in the country exceeded \code{case_threshold}. If both \code{first_date} and \code{case_threshold}
are provided, \code{first_date} will be used as the start of the plot.
}
\examples{

# Puling the HIT-COVID database
hit_data <- hit_pull(add_first_case = FALSE)

# Plotting the case counts for India
intervention_epi(hit_data, country = "IND")

# Starting when there were more than 100 cases
intervention_epi(hit_data, country = "IND", case_threshold = 100)

#Plotting the case counts for the New Zealand from February to May 
intervention_epi(hit_data, country = "NZL", first_date = "3/1/2020", last_date = "5/31/2020")
 
}
\references{
Sam Abbott, Katharine Sherratt, Jonnie Bevan, Hamish Gibbs, Joel Hellewell, James Munday,
Paul Campbell and Sebastian Funk (2020). covidregionaldata: Subnational Data for the
Covid-19 Outbreak. R package version 0.6.0.
https://CRAN.R-project.org/package=covidregionaldata
}
\seealso{
\link{hit_filter}
}
