% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_lookup.R
\docType{data}
\name{intervention_lookup}
\alias{intervention_lookup}
\title{Intervention Look-up Table}
\format{
A data frame with 41 rows and 7 variables:
\describe{
\item{unique_id}{code for the specific intervention used in the \code{unique_id} column of the
database}
\item{intervention_group}{code for the intervention group used in the \code{intervention_group} column of
the database}
\item{intervention_specific}{code for the specific intervention used in the \code{intervention_specific}
column of the database}
\item{intervention_group_name}{name corresponding to the intervention group (\code{intervention})}
\item{intervention_name}{name corresponding to the specific intervention used in the
\code{intervention_name} column of the database}

}
}
\usage{
intervention_lookup
}
\description{
A dataset with the various codes and names for the intervention groups and specific
interventions used in the HIT-COVID database. Can be used to determine interventions
to be used in filtering the database.
}
\keyword{datasets}
