% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_map.R
\name{intervention_map}
\alias{intervention_map}
\title{Mapping of interventions}
\usage{
intervention_map(
  hit_data,
  time_point = Sys.Date(),
  intervention_group = NULL,
  date_format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{hit_data}{the full HIT-COVID database pulled from GitHub, pulled using \link{hit_pull}}

\item{time_point}{character string indicating the desired mapping date (default is Sys.Date())}

\item{intervention_group}{vector of intervention group to filter the data to
(see \link{intervention_lookup} column "intervention_group" or run \link{list_interventions} for options)}

\item{date_format}{character string indicating the desired format of date. (default is "\%m/\%d/\%Y")}
}
\value{
A world map depicting interventions of countries and administrative divisions by a specific date.
}
\description{
This function maps the interventions at both national and administrative level at a specific time.
The HIT-COVID dataset is loaded with \link{hit_pull}, and data of one specific intervention
is pulled using \link{hit_filter}.
}
\details{
The function aims to draw a world map depicting the latest interventions before this \code{time_point}.
If no value is entered for \code{time_point}, it will return a map of most recent interventions.
Both national level and administrative level data are mapped on the same plot.
National level data are represented by the color of country territories, and administrative level data are
represented by dots, both types of data are coded in red (Strongly implemented),
yellow (partially implemented) or green (implementation suspended). The strongest implementation
intensity is used for mapping if for one territory there are multiple entries in the same intervention group on the same day.
}
\examples{

# Puling the HIT-COVID database
hit_data <- hit_pull(add_first_case = FALSE)
 
# Mapping of most recent interventions in "quarantine and isolation" domain
intervention_map(hit_data, intervention_group = "quar_iso")

# Mapping of latest "quarantine and isolation" interventions before 5/24/2020
intervention_map(hit_data, intervention_group = "quar_iso", time_point = "5/24/2020")

}
\references{
https://gadm.org/download_world.html
}
\seealso{
\link{hit_filter}
}
