% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_timeline.R
\name{intervention_timeline}
\alias{intervention_timeline}
\title{Plots a time-line of the intervention data in HIT-COVID}
\usage{
intervention_timeline(
  hit_data,
  facet_by = c("none", "continent", "country", "admin1"),
  first_case_line = TRUE,
  first_death_line = TRUE,
  source = c("WHO", "ECDC"),
  verbose = TRUE,
  ...
)
}
\value{
A plot
}
\description{
This function pulls the HIT-COVID database from the GitHub repo:
\url{https://github.com/HopkinsIDD/hit-covid}. If desired, it also uses the
\code{covidregionaldata} package to add the date of the first case of COVID-19 and the date of
the first death from COVID-19. The source is either the European Centre for Disease Control
(ECDC) or the World Health Organization (WHO).
}
\examples{

\donttest{
#Pulling the HIT-COVID database and adding date of first case and death from WHO
hit_data <- hit_pull(add_first_case = TRUE, source = "WHO")

#Pulling the HIT-COVID databasea and not adding the date of first case and first death
hit_data <- hit_pull(add_first_case = FALSE)
}

}
\references{
Sam Abbott, Katharine Sherratt, Jonnie Bevan, Hamish Gibbs, Joel Hellewell, James Munday,
Paul Campbell and Sebastian Funk (2020). covidregionaldata: Subnational Data for the
Covid-19 Outbreak. R package version 0.6.0.
https://CRAN.R-project.org/package=covidregionaldata
}
\seealso{
\link{hit_filter}, \link{get_first_case}, \link{get_national_data}
}
