% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pulls HIT-COVID database from GitHub}
\usage{
pull_data()
}
\value{
A dataframe with columns as described in the GitHub README
(\url{https://github.com/HopkinsIDD/hit-covid})
}
\description{
This function pulls the HIT-COVID database from the GitHub repo:
\url{https://github.com/HopkinsIDD/hit-covid}.
}
\seealso{
\link{filter_data}
}
