% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_location.R
\docType{data}
\name{admin_location}
\alias{admin_location}
\title{Dataset name}
\format{
A data frame with xxx rows and xxx variables:
\describe{
\item{column name}{column description}
}
}
\usage{
admin_location
}
\description{
Dataset description
}
\references{
\url{https://gadm.org/download_world.html}
}
\keyword{datasets}
