% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_filter.R
\name{hit_pull}
\alias{hit_pull}
\title{Pulls HIT-COVID database from GitHub}
\usage{
hit_pull()
}
\value{
A dataframe with columns as described in the GitHub README
(\url{https://github.com/HopkinsIDD/hit-covid})
}
\description{
This function pulls the HIT-COVID database from the GitHub repo:
\url{https://github.com/HopkinsIDD/hit-covid}.
}
\seealso{
\link{hit_filter}
}
