% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_map.R
\name{intervention_map}
\alias{intervention_map}
\title{Plots a world map of an intervention}
\usage{
intervention_map(
  hit_data,
  time_point = Sys.Date(),
  intervention_group = NULL,
  date_format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{hit_data}{the full HIT-COVID database pulled from GitHub, pulled using \link{hit_pull}}

\item{time_point}{character string indicating the desired mapping date (default is Sys.Date())}

\item{intervention_group}{vector of intervention group to filter the data to
(see \link{intervention_lookup} column "intervention_group" or run \link{get_interventions} for options)}

\item{date_format}{character string indicating the date format of \code{time_point}
(default is "\%m/\%d/\%Y").}
}
\description{
This function maps the status of a specified intervention at a specified date at both national
and administrative level.
}
\details{
The HIT-COVID database must first be loaded with \link{hit_pull} and fed into this function.
The function draws a world map depicting the latest intervention status before the provided
\code{time_point}. If no value is entered for \code{time_point}, it will return a map of most
current intervention status. Both national level and administrative level data are mapped.
National level intervention data are represented by the color of the country polygons.
Administrative level intervention data are represented by circles. Both types of data are coded
in red (Strongly implemented), yellow (partially implemented) or gray (implementation suspended).
The strongest implementation intensity is used for mapping if for one territory there are
multiple entries in the same intervention group on the same day.
}
\examples{

# Puling the HIT-COVID database
hit_data <- hit_pull(add_first_case = FALSE)
 
# Mapping of most recent interventions in "quarantine and isolation" domain
intervention_map(hit_data, intervention_group = "quar_iso")

# Mapping of latest "quarantine and isolation" interventions before 5/24/2020
intervention_map(hit_data, intervention_group = "quar_iso", time_point = "5/24/2020")

}
\references{
https://gadm.org/download_world.html
}
\seealso{
\link{hit_filter}
}
