% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_map.R
\name{intervention_map}
\alias{intervention_map}
\title{Sentence title}
\usage{
intervention_map(
  hit_data,
  time_point = Sys.Date(),
  intervention_group = NULL,
  date.format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{argument_name}{description of the argument}
}
\value{
What the function returns (here a plot)
}
\description{
Paragraph description (comes before arguments in output)
}
\details{
Multiple paragraphs of details as necessary to explain arguments and functioning in more detail
(comes after arguments in output)
}
\examples{
# Add examples of usage showing what different options do

}
\references{
Add any references if relevant
}
\seealso{
\link{hit_filter}
}
