% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pulls HIT-COVID database from GitHub}
\usage{
pull_data(
  country = NULL,
  admin1 = NULL,
  locality = NULL,
  usa_county = NULL,
  intervention = NULL,
  intervention_specific = NULL,
  remove_columns = TRUE
)
}
\arguments{
\item{country}{vector of ISO 3166-1 alpha-3 country codes to filter the data to
(see \link{geo_lookup} for concordance of country codes to names)}

\item{admin1}{vector of the first administrative unit codes to filter the data to
(see \link{geo_lookup} for concordance of admin 1 codes to names)}

\item{locality}{vector of the names of localities to include (this is a free text field)}

\item{usa_county}{logical indicating if the data should be filtered to USA county-level data.
If set to TRUE, only data from \code{country = "USA"} will be provided. If FALSE, all USA
county-level will be removed and country filtering will be determined by the \code{country}
argument.}

\item{intervention}{vector of intervention group codes to filter the data to
(see \link{intervention_lookup} column "intervention" for options)}

\item{intervention_specific}{vector of specific interventions to filter the data to
(see \link{intervention_lookup} column "intervention_specific" for options)}

\item{remove_columns}{a logical indicating if columns with only missing values should be removed
(default is TRUE)}
}
\value{
A dataframe with columns as described in the GitHub README
(\url{https://github.com/HopkinsIDD/hit-covid}) but excluding any columns that have only missing
values if \code{remove_columns = TRUE}
}
\description{
This function pulls the HIT-COVID database from the GitHub repo:
\url{https://github.com/HopkinsIDD/hit-covid}.
If no arguments are supplied the entire database is returned. There is also the option to
filter by country, admin 1 unit, locality, intervention group, and/or specific intervention.
}
\details{
All filtering arguments are optional. If none are provided, the entire database will be returned.
Any or all of the arguments can be specified allowing filtering by location, intervention type
or both. The locality field is used infrequently in this database and this filtering argument
should only be used if it is known that the database has a certain locality (lower than the
admin1 level).

As part of the larger effort, there was a special project to collect some USA county-level data.
If interested in that project, set \code{usa_county} to TRUE. If you want to exclude the USA
county-level data, set \code{usa_county} to FALSE.
}
